more quickly determine recent_rtt
fix up rtt buckets

+++ net_io.h
@@ -130,12 +130,12 @@ struct client
     uint64_t positionCounter; // counter for incoming data
     uint64_t garbage; // amount of garbage we have received from this client
     uint64_t rejected_delayed;
-    double rtt; // last reported rtt in seconds
+    uint32_t rtt; // last reported rtt in milliseconds
     // crude IIR pseudo rolling average, old value factor 0.997
     // cumulative weigth of last 100 packets is 0.26
     // cumulative weigth of last 300 packets is 0.59
     // cumulative weigth of last 600 packets is 0.84
-    double recent_rtt;
+    double recent_rtt; // in milliseconds
     struct epoll_event epollEvent;
     struct net_connector *con;
     int8_t noTimestamps;
